Change the 'adding columns' part of the lesson to follow M+ M- M+ format
This change makes it easier to learn as students are not first presented with an incorrect example. It still allows for understanding the error as the negative model (M-) comes second, then the positive answer (M+) is re-enforced at the end as it is stored while overwriting 'cats'. It also has the advantage of not teaching bad data handling practices ("if the data set doesn't work, change the data set till the technique works").